refactor(recorder): retire the tar write path
Nothing needed `--rez-version 2` any more. `record` writes v3, and since
#1074 `combine`, `filter`, `annotate` and `parquet upgrade` convert a tar
archive rather than producing one -- so the flag selected a container
whose only remaining purpose was to be converted away from.

The flag is REMOVED rather than accepted-and-ignored. Silently honouring
`--rez-version 2` with a v3 archive would be the worst outcome: someone
asking for the tar container would get something else and no indication
their request was dropped. A script still passing it now fails.

`SealPolicy` and `SegmentAccount` move to `recorder::seal_policy`. They
outlived the writer they were born in: "is this open segment due?" and the
byte/row counts that question is asked against are properties of
segmenting a recording, not of the container it lands in, which is why the
v3 writer and hindsight's buffer already used them. `stagger_bucket` and
its determinism test go with them.

What remains of `rez_stream.rs` is now `#[cfg(test)]` -- a fixture builder
and nothing else. It has to stay: proving that v1/v2 archives still READ
requires being able to construct one, and this is what constructs them.

`RezStream` collapses from an enum over containers to a newtype. It still
earns its keep for the two things the recording loop needs and the writer
has no opinion about: what to tell the user after a mid-recording failure,
and how to leave nothing behind when a run captured no samples.
