+++ README.md
@@ -233,11 +233,10 @@ sampling interval:
 rezolus parquet metadata -i out.rez   # reports "not cleanly finalized"
 ```
 
-Pass `--rez-version 2` to write the previous tar container instead, which stages
-at `<output>.partial` and is recoverable only up to its last checkpoint. There
-is no longer a reason to reach for it: `parquet combine`, `filter` and
-`annotate` all read and rewrite both containers, so the default needs no
-opt-out to stay workable.
+The previous tar container is no longer written. Archives recorded by older
+releases still open everywhere, and `rezolus parquet upgrade old.rez` converts
+one to the current container — as does rewriting it with `combine`, `filter` or
+`annotate`, all of which read either container and emit the current one.
 
 ### Viewer
