+++ src/hindsight/mod.rs
@@ -228,7 +228,7 @@ pub fn run(config: Config) {
     // writer's 900 rows is a segment per ~15 minutes at the default 1 s
     // interval, which a faster buffer wants smaller. Everything else about the
     // seal policy — the byte cap and the age cap — stays the writer's.
-    let mut policy = crate::recorder::rez_stream::SealPolicy::default();
+    let mut policy = crate::recorder::seal_policy::SealPolicy::default();
     if let Some(rows) = config.general().segment_rows() {
         // `max(1)`: a zero row target would seal a one-row segment every tick
         // forever rather than doing anything useful with the 0.
